requirements.txt
streamlit==1.32.0
anthropic==0.25.0
langgraph==0.1.4
langchain-anthropic==0.1.3
langchain-core==0.1.46
langchain-community==0.0.38
langchain==0.1.16
python-dotenv==1.0.1
pyspark==3.5.1
delta-spark==3.1.0
faiss-cpu>=1.12.0
sentence-transformers==2.6.1
pandas==2.2.1

